requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djoser==2.1.0
pandas==1.4.3
openpyxl==3.0.10